Revert "Type consistency fixes and Cortex-M0+ assembly optimizations" - requires fixes relative to main before merging.

+++ src/pulse.c
@@ -1,20 +1,18 @@
-#include <stddef.h>
-
+#include "pulse.h"
 #include "board_def.h"
 #include "driver_PORT.h"
 #include "driver_TIME.h"
 #include "emon32.h"
 #include "emon32_samd.h"
-#include "pulse.h"
 
 typedef enum PulseLvl_ { PULSE_LVL_LOW, PULSE_LVL_HIGH } PulseLvl_t;
 
 static uint8_t    pinValue[NUM_OPA];
-static uint32_t   pulseCount[NUM_OPA];
+static uint64_t   pulseCount[NUM_OPA];
 static PulseCfg_t pulseCfg[NUM_OPA];
 static PulseLvl_t pulseLvlLast[NUM_OPA];
 
-PulseCfg_t *pulseGetCfg(const size_t index) {
+PulseCfg_t *pulseGetCfg(const uint32_t index) {
   /* If no pulse counters attached or index out of range, return 0 */
   if ((0 == NUM_OPA) || (index > (NUM_OPA - 1u))) {
     return 0;
@@ -23,10 +21,10 @@ PulseCfg_t *pulseGetCfg(const size_t index) {
   return &pulseCfg[index];
 }
 
-void pulseInit(const size_t index) {
-  const uint8_t opaPUs[] = {PIN_OPA1_PU, PIN_OPA2_PU};
+void pulseInit(const uint32_t index) {
+  const uint_fast8_t opaPUs[] = {PIN_OPA1_PU, PIN_OPA2_PU};
 
-  const uint8_t pin = pulseCfg[index].pin;
+  const uint32_t pin = pulseCfg[index].pin;
 
   /* Enable pull up if configured and allow a delay to charge RC */
   if (pulseCfg[index].puEn) {
@@ -47,23 +45,23 @@ void pulseInit(const size_t index) {
   pulseLvlLast[index] = (PulseLvl_t)pinValue[index];
 }
 
-void pulseSetCount(const size_t index, const uint32_t value) {
+void pulseSetCount(const uint32_t index, const uint64_t value) {
   pulseCount[index] = value;
 }
 
-uint32_t pulseGetCount(const size_t index) { return pulseCount[index]; }
+uint64_t pulseGetCount(const uint32_t index) { return pulseCount[index]; }
 
 void pulseUpdate(void) {
   uint32_t   mask;
   PulseLvl_t level;
 
-  for (size_t i = 0; i < NUM_OPA; i++) {
+  for (uint32_t i = 0; i < NUM_OPA; i++) {
     if (pulseCfg[i].active) {
       mask  = (1 << pulseCfg[i].periods) - 1u;
       level = pulseLvlLast[i];
 
       pinValue[i] <<= 1;
-      pinValue[i] += (uint8_t)portPinValue(pulseCfg[i].grp, pulseCfg[i].pin);
+      pinValue[i] += (uint32_t)portPinValue(pulseCfg[i].grp, pulseCfg[i].pin);
 
       if (0 == (pinValue[i] & mask)) {
         level = PULSE_LVL_LOW;
